Remove Excel_Modell.py due to code redundancy
The Excel_Modell.py file in the flixOpt_excel/Initiation directory has been deleted due to duplicated functionality in other sections of the codebase.
This change is part of a larger refactoring effort to improve the structure and design of the project. Creation of the Energy system from the excel file is now based on classes and functions.